switch to TypeScript & new build 🦁
