+++ package.json
@@ -1,9 +1,11 @@
 {
   "name": "hotstuff",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "A small package to provide an easy way to add Hotkeys easily to any element.",
-  "main": "dist/HotStuff.js",
-  "keywords" : [
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.esm.js",
+  "typings": "types/index.d.ts",
+  "keywords": [
     "keyboard",
     "events",
     "keyboard events",
@@ -13,7 +15,8 @@
     "shortcuts"
   ],
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "rollup -c ./build/rollup.js"
   },
   "repository": {
     "type": "git",
@@ -32,13 +35,7 @@
   },
   "homepage": "https://github.com/laquasicinque/hotstuff#readme",
   "devDependencies": {
-    "babel-preset-es2015": "^6.24.1",
-    "brackets-eslint": "^3.0.6",
-    "gulp": "^3.9.1",
-    "gulp-babel": "^6.1.2",
-    "gulp-eslint": "^4.0.0",
-    "gulp-rename": "^1.2.2",
-    "gulp-uglify": "^3.0.0",
-    "pump": "^1.0.2"
+    "rollup-plugin-typescript2": "^0.19.2",
+    "typescript": "^3.3.3333"
   }
 }
