+++ tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "compilerOptions": {
+    "declaration": true,
+    "declarationDir": "./types/",
+    "target": "es5",
+    "lib": ["es2015", "es2017", "dom"]
+  },
+  "include": [
+    "src/**/*"
+  ]
+}
