+++ types/index.d.ts
@@ -1,0 +1,2 @@
+declare const HotStuff: (el: any) => void;
+export default HotStuff;
