+++ types/keyAliases.d.ts
@@ -1,0 +1,58 @@
+export declare const KEY_ALIAS_MAP: {
+    "+": string;
+    "-": string;
+    ",": string;
+    " ": string;
+    "   ": string;
+    arrowup: string;
+    arrowdown: string;
+    arrowleft: string;
+    arrowright: string;
+    control: string;
+    delete: string;
+};
+export declare const KEYCODE_ALIAS_MAP: {
+    8: string;
+    9: string;
+    13: string;
+    16: string;
+    17: string;
+    18: string;
+    19: string;
+    20: string;
+    27: string;
+    32: string;
+    33: string;
+    34: string;
+    35: string;
+    36: string;
+    37: string;
+    38: string;
+    39: string;
+    40: string;
+    45: string;
+    46: string;
+    59: string;
+    61: string;
+    91: string;
+    92: string;
+    93: string;
+    106: string;
+    107: string;
+    109: string;
+    110: string;
+    111: string;
+    144: string;
+    145: string;
+    186: string;
+    187: string;
+    189: string;
+    190: string;
+    191: string;
+    192: string;
+    219: string;
+    220: string;
+    221: string;
+    222: string;
+    223: string;
+};
